requirements.txt
gradio>=4.0
numpy
Pillow
scikit-image
torch
torchvision
huggingface_hub
pandas
